feat: Implement background color change based on screen size
In the index.html file, I added an h1 title and included the external JavaScript script.js with the "defer" attribute. The h1 title describes the purpose of the page, which is to change the background color based on the screen size.

In the script.js file, I created the myChangeBackgroundColor() function that adjusts the background color of the <html> tag based on the screen width. Three scenarios are considered based on the screen width.

The function is also called when the page loads and responds to screen size changes through an event listener.

This commit adds the feature of changing the background color based on screen size to the page.