feat: Implement count increment functionality
In the job-02 directory, I created an index.html file with a basic HTML structure and a script.js file. In the HTML, I added a paragraph with the ID "count-displayer" and a button with the ID "add-count-btn." I imported the script into the HTML.

In the script.js file, I implemented the myAddCount() function, which increments a counter inside the paragraph. This function is triggered when the previously created button is clicked.

This commit adds the feature to increment the counter when the button is clicked.